+++ src/Agents.jl
@@ -2,6 +2,7 @@ module Agents
 
 using Requires
 using Distributed
+using DataStructures
 using LightGraphs
 using DataFrames
 using Random
@@ -30,6 +31,9 @@ include("simulations/collect.jl")
 include("simulations/paramscan.jl")
 include("simulations/sample.jl")
 
+# Other advanced features
+include("pathfinding/grid_pathfinder.jl")
+
 function __init__()
     # Plot recipes
     @require Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80" begin
@@ -53,7 +57,7 @@ export Models
 
 # Update message:
 display_update = true
-version_number = "4.1"
+version_number = "4.2"
 update_name = "update_v$(version_number)"
 
 if display_update
@@ -63,9 +67,8 @@ if display_update
             """
             \nUpdate message: Agents v$(version_number)
 
-            `AgentBasedModel` now explicitly includes a random number generator, enabling
-            reproducible ABM simulations with Agents.jl.
-            Access it with `model.rng` and seed it with `seed!(model, seed)`!
+            We now have full support for pathfinding, using the A* algorithm, in `GridSpace`!
+            See the "Path-finding" section in the online docs!
             """;
             color = :light_magenta,
         )