+++ src/core/agents.jl
@@ -4,7 +4,7 @@ export AbstractAgent, @agent, GraphAgent, GridAgent, ContinuousAgent, OSMAgent
     YourAgentType <: AbstractAgent
 Agents participating in Agents.jl simulations are instances of user-defined Types that
 are subtypes of `AbstractAgent`. It is almost always the case that mutable Types make
-for a simpler modellign experience.
+for a simpler modelling experience.
 
 Your agent type(s) **must have** the `id` field as first field.
 Depending on the space structure there might be a `pos` field of appropriate type
@@ -38,7 +38,7 @@ abstract type AbstractAgent end
 Create a struct for your agents which includes the mandatory fields required to operate
 in a particular space. Depending on the space of your model, the `AgentSupertype` is
 chosen appropriately from [`GraphAgent`](@ref), [`GridAgent`](@ref),
-[`ContinuousAgent`](@ref).
+[`ContinuousAgent`](@ref), [`OSMAgent`](@ref).
 
 ## Example
 Using
@@ -82,7 +82,7 @@ end
 """
     GridAgent{D}
 Combine with [`@agent`](@ref) to create an agent type for `D`-dimensional
-[`GraphSpace`](@ref). It attributes the fields `id::Int, pos::NTuple{D,Int}`
+[`GridSpace`](@ref). It attributes the fields `id::Int, pos::NTuple{D,Int}`
 to the start of the agent type.
 """
 mutable struct GridAgent{D} <: AbstractAgent